+++ src/Request.php
@@ -2,17 +2,21 @@
 
 namespace MetaRush\Perm;
 
+use Psr\Http\Message\ServerRequestInterface;
+
 class Request
 {
     private int $userId;
     private int $roleId;
     private string $resourceId;
+    private ?ServerRequestInterface $serverRequest;
 
-    public function __construct(int $userId, int $roleId, string $resourceId)
+    public function __construct(int $userId, int $roleId, string $resourceId, ?ServerRequestInterface $serverRequest = null)
     {
         $this->userId = $userId;
         $this->roleId = $roleId;
         $this->resourceId = $resourceId;
+        $this->serverRequest = $serverRequest;
     }
 
     public function getUserId(): int
@@ -28,6 +32,11 @@ class Request
     public function getResourceId(): string
     {
         return $this->resourceId;
+    }
+
+    public function getServerRequest(): ?ServerRequestInterface
+    {
+        return $this->serverRequest;
     }
 
 }